requirements.txt
Flask
flask-cors
Pillow
imagehash
gunicorn
tensorflow
scikit-learn
numpy
